% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/fastrank.R
\docType{package}
\name{fastrank-package}
\alias{fastrank-package}
\title{Ranking vectors with low overhead}
\description{
An R package providing fast ranking for integer and numeric vectors, as an
alternative to calling .Internal(rank(...)), which packages cannot do.
}
\details{
The motivation for this comes from my development of the `nestedRanksTest`
package.  A standard run with the default 10,000 bootstrap iterations takes
a few seconds to complete on a test data set.
}
\references{
Source?
R-core?

\url{https://github.com/douglasgscofield/fastrank}
}

