% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/fastrank.R
\name{fastrank}
\alias{fastrank}
\title{Rank integer and numeric vectors with low overhead}
\usage{
fastrank(x)
}
\arguments{
\item{x}{A vector of values to rank.}

\item{ties.method}{Method for resolving rank ties in \code{x}.  Provides
                    all available in \code{rank}.  This does **not** provide
                    method \code{"random"}.}
}
\value{
An integert vector of ranks of values in \code{x}, with length
        the same as \code{length(x)}.
}
\description{
An R function providing fast ranking for integer and numeric vectors, as an
alternative to calling .Internal(rank(...)), which packages cannot do.
}
\details{
There is still a bit of overhead in calling this method.  If you really want
the fastest interface, then call the C function directly.
}
\note{
The vector must not include NAs or NaNs.  This is **not** checked.
}
\references{
\url{https://github.com/douglasgscofield/fastrank}
}
\seealso{
\code{\link{rank}}, \code{\link{order}},
         \code{\link{sort}}
}
\keyword{internal}

