% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/fastrank.R
\name{fastrank_average}
\alias{fastrank_average}
\title{Rank logical, integer, numeric, complex vectors assigning ties the average rank}
\usage{
fastrank_average(x)
}
\arguments{
\item{x}{A vector of numeric (double) values to rank}
}
\value{
A numeric vector of ranks of values in \code{x}, with length
the same as \code{length(x)}.  Ties are broken by giving entries the
average rank of the tied entries.
}
\description{
An R function providing fast ranking for numeric vectors, assigning tied
values the average rank.
}
\details{
There is still a bit of overhead in calling this method.  If you really want
the fastest interface, then call .\code{.Internal(rank(...))} function
directly.
}
\note{
The vector must not include NAs or NaNs.  This is **not** checked.
}
\references{
\url{https://github.com/douglasgscofield/fastrank}
}
\seealso{
\code{\link{rank}}
}
\keyword{internal}

