% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/fastrank.R
\name{fastrank_num_avg}
\alias{fastrank_num_avg}
\title{Rank integer and numeric vectors with low overhead}
\usage{
fastrank_num_avg(x)
}
\arguments{
\item{x}{A vector of values to rank.}
}
\value{
An integer vector of ranks of values in \code{x}, with length
the same as \code{length(x)}.  Ties are broken by giving entries the
average rank of the tied entries.
}
\description{
An R function providing fast ranking for integer and numeric vectors, as an
alternative to calling .Internal(rank(...)), which packages cannot do.
}
\details{
There is still a bit of overhead in calling this method.  If you really want
the fastest interface, then call the C function directly.
}
\note{
The vector must not include NAs or NaNs.  This is **not** checked.
}
\references{
\url{https://github.com/douglasgscofield/fastrank}
}
\seealso{
\code{\link{rank}}
}
\keyword{internal}

