% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/fastrank.R
\docType{data}
\name{fastrank_numeric_average}
\alias{fastrank_numeric_average}
\title{Rank integer and numeric vectors with low overhead}
\format{\preformatted{List of 4
 $ name         : chr "fastrank_numeric_average"
 $ address      :Class 'RegisteredNativeSymbol' <externalptr> 
 $ package      :List of 5
  ..$ name         : chr "fastrank"
  ..$ path         : chr "/Users/douglasgscofield/Dropbox/_my-R-packages/fastrank/src/fastrank.so"
  ..$ dynamicLookup: logi TRUE
  ..$ handle       :Class 'DLLHandle' <externalptr> 
  ..$ info         :Class 'DLLInfoReference' <externalptr> 
  ..- attr(*, "class")= chr "DLLInfo"
 $ numParameters: int 1
 - attr(*, "class")= chr [1:2] "CRoutine" "NativeSymbolInfo"
}}
\usage{
fastrank_numeric_average
}
\arguments{
\item{x}{A vector of values to rank.}
}
\value{
An integer vector of ranks of values in \code{x}, with length
the same as \code{length(x)}.  Ties are broken by giving entries the
average rank of the tied entries.
}
\description{
An R function providing fast ranking for integer and numeric vectors, as an
alternative to calling .Internal(rank(...)), which packages cannot do.
}
\details{
There is still a bit of overhead in calling this method.  If you really want
the fastest interface, then call the C function directly.
}
\note{
The vector must not include NAs or NaNs.  This is **not** checked.
}
\references{
\url{https://github.com/douglasgscofield/fastrank}
}
\seealso{
\code{\link{rank}}
}
\keyword{internal}

