% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fastrank.R
\docType{package}
\name{fastrank-package}
\alias{fastrank-package}
\title{Ranking vectors with low overhead}
\description{
An R package providing fast ranking for integer and numeric vectors, as an
alternative to calling .Internal(rank(...)), which packages cannot do.
}
\details{
The motivation for this comes from my development of the
\code{nestedRanksTest} package.  A standard run with the default 10,000
bootstrap iterations takes a few seconds to complete on a test data set.
Profiling with \code{library(lineprof)} revealed that the bottleneck here is
in the utility function \code{nestedRanksTest_Z}, specifically in the
calculation of ranks via the base R function \code{rank}.  A stripped-down
\code{rank_new} is 8-9 times faster than the default rank for a vector of 100
values.  For 1000-value vectors the speedup is more modest, about 2&times;,
and for 10,000-value vectors the speedup is only in the neighbourhood of
20-30%.
}
\references{
\url{https://github.com/douglasgscofield/fastrank}
}

