% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/fastrank.R
\name{fastrank}
\alias{fastrank}
\title{Rank vectors with low overhead}
\usage{
fastrank(x, ties.method = "average")
}
\arguments{
\item{x}{A vector of values to rank.  Note that character vectors are not
accepted, as the internal R routines for comparing characters as R does
are not part of the R API.  Please use \code{\link{rank}} from base R
to rank character vectors.}

\item{ties.method}{Method for resolving rank ties in \code{x}, all in
\code{\link{rank}} are available}

\item{x}{Vector to calculate ranks for}

\item{find}{Method for finding \code{ties.method}, either 1 or 2}
}
\value{
A vector of ranks of values in \code{x}, with length
the same as \code{length(x)}.  Ranks of tied values are handled according
to \code{ties.method}, see \code{\link{rank}}.  When \code{ties.method} is
\code{"average"}, a numeric vector is returned, otherwise an integer
vector is returned.
}
\description{
An R function providing fast ranking vectors, as an alternative to calling
\code{.Internal(rank(...))}, which packages cannot do.  This entry is as
general as it can be, but so far there is still a bit of overhead in
calling this method.  If you really want the fastest interface, try using
one of the other entry points (e.g., \code{fastrank_numeric_average}) or
(if you are not writing package code) call the C function directly.
}
\note{
The vector must not include NAs or NaNs.  This is **not** checked.

The vector must also not be of type character.
}
\references{
\url{https://github.com/douglasgscofield/fastrank}
}
\seealso{
\code{\link{rank}}
}
\keyword{internal}

